% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dep_reactdom}
\alias{dep_reactdom}
\title{Internal helper to load the React DOM Model as an
\code{htmltools::htmlDependency} for your widget.}
\usage{
dep_reactdom()
}
\value{
An \link[htmltools:htmlDependency]{htmltools::htmlDependency} for SurveyJS v2.2.6
}
\description{
Internal helper to load the React DOM Model as an
\code{htmltools::htmlDependency} for your widget.
}
\keyword{internal}
