% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dep_surveyjs_core}
\alias{dep_surveyjs_core}
\title{SurveyJS core HTML dependency}
\usage{
dep_surveyjs_core()
}
\value{
An \link[htmltools:htmlDependency]{htmltools::htmlDependency} for SurveyJS v2.2.6
}
\description{
Internal helper that injects the SurveyJS core library as an
\code{htmltools::htmlDependency} for your widget.
}
\keyword{internal}
