% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyjs.R
\name{surveyjs-shiny}
\alias{surveyjs-shiny}
\alias{surveyjsOutput}
\alias{renderSurveyjs}
\title{Shiny bindings for the SurveyJS widget}
\usage{
surveyjsOutput(outputId, width = "100\%", height = "400px")

renderSurveyjs(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{Output variable to read from.}

\item{width, height}{Size of the output container (CSS unit or number).}

\item{expr}{An expression that generates a surveyjs widget.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression? (With \code{quote()}.)}
}
\description{
Shiny bindings for the SurveyJS widget
}
