% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyjs.R
\name{surveyjs}
\alias{surveyjs}
\title{SurveyJS widget for Shiny with support for JavaScript hooks}
\usage{
surveyjs(
  schema,
  data = NULL,
  readOnly = FALSE,
  live = FALSE,
  theme = "DefaultLight",
  theme_vars = NULL,
  locale = NULL,
  pre_render_hook = NULL,
  post_render_hook = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{schema}{List or JSON string; must follow
\href{https://surveyjs.io/form-library/documentation/json-schema}{SurveyJS JSON Schema}.}

\item{data}{Initial values.}

\item{readOnly}{Render in read-only mode.}

\item{live}{Live update responses?}

\item{theme}{Theme name.}

\item{theme_vars}{Named list of CSS variables (e.g. \code{--sjs-primary-backcolor}).}

\item{locale}{Language code (e.g. \code{"en"}, \code{"de"})}

\item{pre_render_hook}{JavaScript code as a string to run before render.}

\item{post_render_hook}{JavaScript code as a string to run after render.}

\item{width, height}{Optional size specs.}

\item{elementId}{Optional element id for the container.}
}
\description{
Render a \href{https://surveyjs.io/}{SurveyJS} form from a JSON schema.
}
\details{
This widget renders a fully customizable survey using SurveyJS v2 inside a Shiny app.
It supports themes, read-only mode, initial values, localization, and real-time updates.

Final answers are delivered to \code{input[[paste0(id, "_data")]]} after the user
clicks \strong{Complete}. If \code{live = TRUE}, updates are delivered to
\code{input[[paste0(id, "_data_live")]]} during input.
}
