% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyjs_themes.R
\name{surveyjs_themes}
\alias{surveyjs_themes}
\title{Available SurveyJS Themes}
\usage{
surveyjs_themes()
}
\value{
A character vector of theme names.
}
\description{
Common built-in themes for use with the \code{theme} argument in \code{surveyjs()}.
}
\examples{
surveyjs_themes()
}
